Maintainability: Netlify form as a component
